+++ NSString+Base64.m
@@ -17,7 +17,7 @@ static char base64EncodingTable[64] = {
 
 @implementation NSString (Base64Additions)
 
-+ (NSString *)base64StringFromData: (NSData *)data length: (int)length {
++ (NSString *)base64StringFromData: (NSData *)data length: (NSUInteger)length {
   unsigned long ixtext, lentext;
   long ctremaining;
   unsigned char input[3], output[4];
